pyproject.toml
[tool.poetry]
name = "car-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.2.1"
django = "^6.0.1"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"
coverage = "^7.13.2"
drf-yasg = "^1.21.14"
requests = "^2.32.5"
django-cors-headers = "^4.9.0"

[tool.poetry.group.dev.dependencies]

psycopg2-binary = "^2.9.11"
pillow = "^12.1.0"
djangorestframework = "^3.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
